Opponent APIs aded and update dpom for war

+++ src/main/java/edu/sjsu/cmpe275/lab2/controller/SponsorController.java
@@ -1,5 +1,7 @@
 package edu.sjsu.cmpe275.lab2.controller;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -13,7 +15,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import edu.sjsu.cmpe275.lab2.mapper.SponsorResponse;
+import edu.sjsu.cmpe275.lab2.model.Player;
 import edu.sjsu.cmpe275.lab2.model.Sponsor;
+import edu.sjsu.cmpe275.lab2.services.PlayerService;
 import edu.sjsu.cmpe275.lab2.services.SponsorService;
 import edu.sjsu.cmpe275.lab2.validators.GameApisValidator;
 
@@ -23,6 +27,9 @@ public class SponsorController {
 
 	@Autowired
 	SponsorService sponsorService;
+
+	@Autowired
+	PlayerService playerService;
 
 	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<?> addNewSponsor(@RequestBody Sponsor sponsor) {
@@ -87,9 +94,16 @@ public class SponsorController {
 		HttpStatus httpStatus = null;
 		Sponsor sponsor = sponsorService.getSponsor(sponsorId);
 		if (sponsor != null) {
-			sponsorService.deleteSponsor(sponsorId);
-			sponsorResponse.setMsg("Successfully Deleted Sponsor.");
-			httpStatus = HttpStatus.OK;
+			List<Player> players = playerService.getPlayerBySponsor(sponsor);
+			if (players.isEmpty()) {
+				sponsorService.deleteSponsor(sponsorId);
+				sponsorResponse.setMsg("Successfully Deleted Sponsor.");
+				httpStatus = HttpStatus.OK;
+			} else {
+				sponsorResponse
+						.setMsg("Sponsor Exist for some players. Cannot be deleted");
+				httpStatus = HttpStatus.NOT_FOUND;
+			}
 		} else {
 			sponsorResponse.setMsg("Sponsor does not exist");
 			httpStatus = HttpStatus.NOT_FOUND;
@@ -101,7 +115,8 @@ public class SponsorController {
 
 	@PostMapping(path = "/{sponsorId}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<?> updateSponsor(
-			@PathVariable(value = "sponsorId") String sponsorId, @RequestBody Sponsor sponsorRequest) {
+			@PathVariable(value = "sponsorId") String sponsorId,
+			@RequestBody Sponsor sponsorRequest) {
 
 		SponsorResponse sponsorResponse = new SponsorResponse();
 		ResponseEntity res = null;
@@ -110,17 +125,19 @@ public class SponsorController {
 				sponsorRequest, sponsorResponse);
 		if (isValid) {
 			Sponsor sponsor = sponsorService.getSponsor(sponsorId);
-			if(sponsor != null) {
-				Sponsor updatedSponsor = sponsorService.updateSponsor(sponsor, sponsorRequest);
+			if (sponsor != null) {
+
+				Sponsor updatedSponsor = sponsorService.updateSponsor(sponsor,
+						sponsorRequest);
 				sponsorResponse.setSponsor(updatedSponsor);
 				sponsorResponse.setMsg("Sponsor updated Successfully");
 				httpStatus = HttpStatus.OK;
-			}
-			else {				
+
+			} else {
 				sponsorResponse.setMsg("Sponsor does not exist");
-				httpStatus = HttpStatus.NOT_FOUND;				
+				httpStatus = HttpStatus.NOT_FOUND;
 			}
-			
+
 		} else {
 			httpStatus = HttpStatus.BAD_REQUEST;
 		}
